fixed returning throwUnsupportedError and moved helper
